+++ generic-cpp-receiver/src/main/cpp/RCode/commands/RCodeExecutionCommand.cpp
@@ -27,7 +27,7 @@ void RCodeExecutionCommand::execute(RCodeCommandSlot *slot,
         bool fits = false;
         executionSpaceAddress_t targetPos = 0;
         int fieldSectionNum = slot->getFields()->countFieldSections('A');
-        int effectiveSize = (fieldSectionNum - 1) * sizeof(fieldUnit);
+        uint16_t effectiveSize = (fieldSectionNum - 1) * sizeof(fieldUnit);
         fieldUnit first = slot->getFields()->get('A', 0);
         while (first != 0) {
             first >>= 8;
